+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 ## Fixes
 
 - Use the standard library rather than `curl` for app uploads and handle error responses [#215](https://github.com/bugsnag/maze-runner/pull/215)
+- Expand file paths for `--app` and `--bs-local` options [#216](https://github.com/bugsnag/maze-runner/pull/216)
 
 # 4.5.0 - 2021/02/01